Remove `condition` keyword; refactor
- `Precondition` is now what `PreconditionKind` was before, since it did
not hold any further information than it.
- `Precondition` no longer holds a reason.
