+++ tests/nightly/function/compile_fail/extra_precondition.stderr
@@ -1,9 +1,9 @@
 error[E0308]: mismatched types
  --> $DIR/extra_precondition.rs:8:5
   |
-8 | /     #[assert_pre(condition("must be bar", reason = "is bar"))]
-9 | |     #[assert_pre(condition("must be baz", reason = "is baz"))]
-  | |______________________________________________________________^ expected a tuple with 1 element, found one with 2 elements
+8 | /     #[assert_pre("must be bar", reason = "is bar")]
+9 | |     #[assert_pre("must be baz", reason = "is baz")]
+  | |___________________________________________________^ expected a tuple with 1 element, found one with 2 elements
   |
   = note: expected struct `std::marker::PhantomData<(pre::const_generics_types::CustomConditionHolds<"must be bar">,)>`
              found struct `std::marker::PhantomData<(pre::const_generics_types::CustomConditionHolds<"must be bar">, pre::const_generics_types::CustomConditionHolds<"must be baz">)>`
