+++ tests/nightly/function/compile_fail/precondition_missing.stderr
@@ -1,8 +1,8 @@
 error[E0308]: mismatched types
  --> $DIR/precondition_missing.rs:9:5
   |
-9 |     #[assert_pre(condition("must be bar", reason = "is bar"))]
-  |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ expected a tuple with 2 elements, found one with 1 element
+9 |     #[assert_pre("must be bar", reason = "is bar")]
+  |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ expected a tuple with 2 elements, found one with 1 element
   |
   = note: expected struct `std::marker::PhantomData<(pre::const_generics_types::CustomConditionHolds<"must be bar">, pre::const_generics_types::CustomConditionHolds<"must be baz">)>`
              found struct `std::marker::PhantomData<(pre::const_generics_types::CustomConditionHolds<"must be bar">,)>`
