+++ tests/nightly/function/pass/single_precondition.rs
@@ -1,10 +1,10 @@
 use pre::{check_pre, pre};
 
-#[pre(condition("must be bar"))]
+#[pre("must be bar")]
 fn foo() {}
 
 #[check_pre]
 fn main() {
-    #[assert_pre(condition("must be bar", reason = "is bar"))]
+    #[assert_pre("must be bar", reason = "is bar")]
     foo()
 }
