+++ tests/stable/function/compile_fail/extra_precondition.rs
@@ -1,11 +1,11 @@
 use pre::{check_pre, pre};
 
-#[pre(condition("must be bar"))]
+#[pre("must be bar")]
 fn foo() {}
 
 #[check_pre]
 fn main() {
-    #[assert_pre(condition("must be bar", reason = "is bar"))]
-    #[assert_pre(condition("must be baz", reason = "is baz"))]
+    #[assert_pre("must be bar", reason = "is bar")]
+    #[assert_pre("must be baz", reason = "is baz")]
     foo()
 }
