+++ tests/stable/function/compile_fail/precondition_missing.rs
@@ -1,11 +1,11 @@
 use pre::{check_pre, pre};
 
-#[pre(condition("must be bar"))]
-#[pre(condition("must be baz"))]
+#[pre("must be bar")]
+#[pre("must be baz")]
 fn foo() {}
 
 #[check_pre]
 fn main() {
-    #[assert_pre(condition("must be bar", reason = "is bar"))]
+    #[assert_pre("must be bar", reason = "is bar")]
     foo()
 }
