+++ tests/stable/function/compile_fail/undefined_precondition.rs
@@ -4,6 +4,6 @@ fn foo() {}
 
 #[check_pre]
 fn main() {
-    #[assert_pre(condition("must be bar", reason = "is bar"))]
+    #[assert_pre("must be bar", reason = "is bar")]
     foo()
 }
