+++ tests/stable/function/compile_fail/undefined_precondition.stderr
@@ -10,7 +10,7 @@ error[E0061]: this function takes 0 arguments but 1 argument was supplied
 3 |   fn foo() {}
   |   -------- defined here
 ...
-7 | /     #[assert_pre(condition("must be bar", reason = "is bar"))]
+7 | /     #[assert_pre("must be bar", reason = "is bar")]
 8 | |     foo()
   | |     ^^-
   | |_____|_|
