+++ tests/templates/function/compile_fail/missing_assert.rs
@@ -1,6 +1,6 @@
 use pre::{check_pre, pre};
 
-#[pre(condition("must be bar"))]
+#[pre("must be bar")]
 fn foo() {}
 
 #[check_pre]
