+++ tests/templates/function/compile_fail/missing_reason.rs
@@ -1,10 +1,10 @@
 use pre::{check_pre, pre};
 
-#[pre(condition("must be bar"))]
+#[pre("must be bar")]
 fn foo() {}
 
 #[check_pre]
 fn main() {
-    #[assert_pre(condition("must be bar"))]
+    #[assert_pre("must be bar")]
     foo()
 }
